pyproject.toml
[project]
name = "garagem-fb"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Rafaella", email = "rafaelacamilly2018@gmail.com"},
    {name = "", email = ""},
]
dependencies = [
    "django>=5.0.3",
    "djangorestframework>=3.15.1",
    "django-cors-headers>=4.3.1",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true
